feat: Add shoppable link functionality
Adds support for shoppable product links to blog posts. Users can now optionally add a product URL and name when uploading a new post. This information will be displayed on the PostCard component as a "Shop Now" button, linking to the specified product.

The `Post` type and `UploadModal` have been updated to accommodate these new fields.